requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
pillow>=10.0.0
